fix: move duplicate queries to queries.js
