+++ components/collective-page/graphql/queries.js
@@ -1,6 +1,10 @@
 import { gql } from '@apollo/client';
 
+import { gqlV2 } from '../../../lib/graphql/helpers';
+
 import { MAX_CONTRIBUTORS_PER_CONTRIBUTE_CARD } from '../../contribute-cards/Contribute';
+import { expensesListFieldsFragment } from '../../expenses/graphql/fragments';
+import { transactionsQueryCollectionFragment } from '../../transactions/graphql/fragments';
 
 import * as fragments from './fragments';
 
@@ -273,6 +277,22 @@ export const collectivePageQuery = gql`
   ${fragments.updatesFieldsFragment}
   ${fragments.contributorsFieldsFragment}
   ${fragments.collectiveNavbarFieldsFragment}
+`;
+
+export const budgetSectionQuery = gqlV2/* GraphQL */ `
+  query BudgetSection($slug: String!, $limit: Int!, $kind: [TransactionKind]) {
+    transactions(account: { slug: $slug }, limit: $limit, hasExpense: false, kinds: $kind) {
+      ...TransactionsQueryCollectionFragment
+    }
+    expenses(account: { slug: $slug }, limit: $limit) {
+      totalCount
+      nodes {
+        ...ExpensesListFieldsFragment
+      }
+    }
+  }
+  ${transactionsQueryCollectionFragment}
+  ${expensesListFieldsFragment}
 `;
 
 export const getCollectivePageQueryVariables = slug => {
